// enums/operations.enum.ts
export enum Operations {
  Exit = 'Exit',
  InsertOne = 'Insert one item',
  InsertRandom = 'Insert random items',
  Clear = 'Clear data structure',
  Length = 'Get current size of data structure',
  Delete = 'Remove item from data structure',
  Update = 'Update specific item in the data structure',
  Contains = 'Search for specific item in data structure',
  Traverse = 'Traverse entire data structure'
};